Log an unhandled status on the remaining forges
Every forge fell through to a default that guessed a value silently. The
statuses that legitimately land there are now listed explicitly, so the
default only sees a value outside the enum, and warns.

The gitea and forgejo status tests move into the test file next to their
source.
