+++ server/forge/bitbucket/convert_test.go
@@ -34,6 +34,10 @@ func Test_convertStatus(t *testing.T) {
 	assert.Equal(t, statusFailure, convertStatus(model.StatusFailure))
 	assert.Equal(t, statusFailure, convertStatus(model.StatusKilled))
 	assert.Equal(t, statusFailure, convertStatus(model.StatusError))
+	assert.Equal(t, statusFailure, convertStatus(model.StatusDeclined))
+	assert.Equal(t, statusFailure, convertStatus(model.StatusSkipped))
+	assert.Equal(t, statusFailure, convertStatus(model.StatusCanceled))
+	assert.Equal(t, statusFailure, convertStatus(model.StatusValue("bogus")))
 }
 
 func Test_convertRepo(t *testing.T) {
