+++ server/forge/gitlab/status_test.go
@@ -38,8 +38,10 @@ func TestGetStatus(t *testing.T) {
 		{model.StatusFailure, gitlab.Failed},
 		{model.StatusError, gitlab.Failed},
 		{model.StatusKilled, gitlab.Canceled},
-		// unknown statuses fall back to failed
 		{model.StatusDeclined, gitlab.Failed},
+		{model.StatusSkipped, gitlab.Failed},
+		{model.StatusCanceled, gitlab.Failed},
+		{model.StatusValue("bogus"), gitlab.Failed},
 	}
 
 	for _, tt := range tests {
